pyproject.toml
[build-system]
requires = [ "flit_core" ]
build-backend = "flit_core.buildapi"

[project]
name = "ntt"
version = "0.1.0"
description = "Library for sport analysis"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Romain Vuillemot", email = "romain.vuillemot@ec-lyon.fr"},
]
maintainers = [
    {name = "Romain Vuillemot", email = "romain.vuillemot@ec-lyon.fr"},
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Natural Language :: English",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Multimedia :: Video",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: Software Development :: Libraries"
]
requires-python = ">=3.9"
dependencies = ["requests", "opencv-python", "ffmpeg-python", "moviepy", "scipy", "python-dotenv"]

[project.urls]
repository = "https://github.com/centralelyon/ntt/"

[project.optional-dependencies]
dev = [
    "black",
    "pytest",
    "flit"
]
